requirements.txt
Flask
requests
python-dotenv
SQLAlchemy
psycopg2-binary
